Updated a LOOOOT of stuff
Added a way of transferring files and seeing what files needed to be transferred, as well as how many. I also changed the example.py to include a better example of code to use, as well as minor adjustments to previous algorithms. New README examples, and showing off some other cool stuff.